# openapi.yml
openapi: 3.0.1
info:
  title: Swagger Ride Lab
  version: 1.0.0
  description: Swagger for geohash API
  contact:
    email: dicodiaz@gmail.com
    name: Dico Diaz
    url: "https://github.com/dicodiaz"
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "http://localhost:3030/api/v1"
tags:
  - name: ride
    description: Operations about ride
  - name: customer
    description: Operations about customer
  - name: driver
    description: Operations about driver
  - name: car
    description: Operations about car
paths:
  /cars:
    get:
      tags:
        - car
      summary: Get all cars
      operationId: getAllCars
      responses:
        default:
          description: successful operation
          content: {}
  /cars/register:
    post:
      tags:
        - car
      summary: Create car
      operationId: createCar
      requestBody:
        description: Created car object
        content:
          "*/*":
            schema:
              $ref: "#/components/schemas/Car"
        required: true
      responses:
        default:
          description: successful operation
          content: {}
      x-codegen-request-body-name: body
  /cars/{id}/detail:
    get:
      tags:
        - car
      summary: Get car by id
      operationId: getCarById
      parameters:
        - name: id
          in: path
          description: The id that needs to be fetched. Use 1 for testing.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/Car"
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
        "400":
          description: Invalid id supplied
          content: {}
        "404":
          description: Car not found
          content: {}
    put:
      tags:
        - car
      summary: Update car
      operationId: updateCar
      parameters:
        - name: id
          in: path
          description: The id that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update car object
        content:
          "*/*":
            schema:
              $ref: "#/components/schemas/Car"
        required: true
      responses:
        "400":
          description: Invalid car supplied
          content: {}
        "404":
          description: Car not found
          content: {}
      x-codegen-request-body-name: body
    delete:
      tags:
        - car
      summary: Delete car
      operationId: deleteCar
      parameters:
        - name: id
          in: path
          description: The id that needs to be to be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Invalid plate supplied
          content: {}
        "404":
          description: Car not found
          content: {}
  /users/drivers:
    get:
      tags:
        - driver
      summary: Get all drivers
      operationId: getAllDrivers
      responses:
        default:
          description: successful operation
          content: {}
  /users/drivers/register:
    post:
      tags:
        - driver
      summary: Create driver
      operationId: createDriver
      requestBody:
        description: Created driver object
        content:
          "*/*":
            schema:
              $ref: "#/components/schemas/Driver"
        required: true
      responses:
        default:
          description: successful operation
          content: {}
      x-codegen-request-body-name: body
  /users/drivers/{id}/detail:
    get:
      tags:
        - driver
      summary: Get driver by id
      operationId: getDriverById
      parameters:
        - name: id
          in: path
          description: The id that needs to be fetched. Use 1 for testing.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/Driver"
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "400":
          description: Invalid id supplied
          content: {}
        "404":
          description: Driver not found
          content: {}
    put:
      tags:
        - driver
      summary: Update driver
      operationId: updateDriver
      parameters:
        - name: id
          in: path
          description: The id that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update driver object
        content:
          "*/*":
            schema:
              $ref: "#/components/schemas/Driver"
        required: true
      responses:
        "400":
          description: Invalid driver supplied
          content: {}
        "404":
          description: Driver not found
          content: {}
      x-codegen-request-body-name: body
    delete:
      tags:
        - driver
      summary: Delete driver
      operationId: deleteDriver
      parameters:
        - name: id
          in: path
          description: The id that needs to be to be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Invalid driverId supplied
          content: {}
        "404":
          description: Driver not found
          content: {}
  /users/customers:
    get:
      tags:
        - customer
      summary: Get all customers
      operationId: getAllCustomers
      responses:
        default:
          description: successful operation
          content: {}
  /users/customers/register:
    post:
      tags:
        - customer
      summary: Create customer
      operationId: createCustomer
      requestBody:
        description: Created customer object
        content:
          "*/*":
            schema:
              $ref: "#/components/schemas/Customer"
        required: true
      responses:
        default:
          description: successful operation
          content: {}
      x-codegen-request-body-name: body
  /users/customers/{id}/detail:
    get:
      tags:
        - customer
      summary: Get customer by id
      operationId: getCustomerById
      parameters:
        - name: id
          in: path
          description: The id that needs to be fetched. Use 1 for testing.
          required: true
          schema:
            type: integer
            # Por qué 64 bits?
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            # Por qué dar la opción de devolver xml?
            application/xml:
              schema:
                $ref: "#/components/schemas/Customer"
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        # La validación de id inválidas no se hace en el front end?
        "400":
          description: Invalid id supplied
          content: {}
        "404":
          description: Customer not found
          content: {}
    put:
      tags:
        - customer
      summary: Update customer
      operationId: updateCustomer
      parameters:
        - name: id
          in: path
          description: The id that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update customer object
        content:
          "*/*":
            schema:
              $ref: "#/components/schemas/Customer"
        required: true
      responses:
        "400":
          description: Invalid customer supplied
          content: {}
        "404":
          description: Customer not found
          content: {}
      x-codegen-request-body-name: body
    delete:
      tags:
        - customer
      summary: Delete customer
      operationId: deleteCustomer
      parameters:
        - name: id
          in: path
          description: The id that needs to be to be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Invalid customerId supplied
          content: {}
        "404":
          description: Customer not found
          content: {}
  /rides:
    get:
      tags:
        - ride
      summary: Get all rides
      operationId: getAllRides
      responses:
        default:
          description: successful operation
          content: {}
  /rides/register:
    post:
      tags:
        - ride
      summary: Create ride
      operationId: createRide
      requestBody:
        description: Created ride object
        content:
          "*/*":
            schema:
              $ref: "#/components/schemas/Ride"
        required: true
      responses:
        default:
          description: successful operation
          content: {}
      x-codegen-request-body-name: body
  /rides/{id}/detail:
    get:
      tags:
        - ride
      summary: Get ride by id
      operationId: getRideById
      parameters:
        - name: id
          in: path
          description: The id that needs to be fetched. Use 1 for testing.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/Ride"
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
        "400":
          description: Invalid id supplied
          content: {}
        "404":
          description: Ride not found
          content: {}
    put:
      tags:
        - ride
      summary: Update ride
      operationId: updateRide
      parameters:
        - name: id
          in: path
          description: The id that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update ride object
        content:
          "*/*":
            schema:
              $ref: "#/components/schemas/Ride"
        required: true
      responses:
        "400":
          description: Invalid ride supplied
          content: {}
        "404":
          description: Ride not found
          content: {}
      x-codegen-request-body-name: body
    delete:
      tags:
        - ride
      summary: Delete ride
      operationId: deleteRide
      parameters:
        - name: id
          in: path
          description: The id that needs to be to be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Invalid driver_id supplied
          content: {}
        "404":
          description: Ride not found
          content: {}
components:
  schemas:
    Car:
      type: object
      properties:
        id:
          type: string
        plate:
          type: string
        model:
          type: string
        motor:
          type: string
        chassis:
          type: string
        color:
          type: string
      xml:
        name: Car
    Driver:
      type: object
      properties:
        id:
          type: string
        driver_id:
          type: integer
          format: int64
        firstname:
          type: string
        lastname:
          type: string
        age:
          type: integer
          format: int32
        username:
          type: string
        phone:
          type: string
        password:
          type: string
      xml:
        name: Driver
    Customer:
      type: object
      properties:
        id:
          type: string
        customer_id:
          type: integer
          format: int64
        firstname:
          type: string
        lastname:
          type: string
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
      xml:
        name: Customer
    Ride:
      type: object
      properties:
        id:
          type: string
        first_point:
          $ref: "#/components/schemas/Point"
        target_point:
          $ref: "#/components/schemas/Point"
        driver_id:
          type: string
        customer_id:
          type: string
        time:
          type: string
      xml:
        name: Ride
    Point:
      type: object
      properties:
        lat:
          type: string
        lng:
          type: string
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: "http://geohashapp.swagger.io/oauth/dialog"
          scopes:
            "write:user": modify user in your account
            "read:user": read your user
    api_key:
      type: apiKey
      name: api_key
      in: header